Email Verification
- Created two helper functions for sending emails.
- Different kinds of tokens are being generated now.
- Sending proper templated emails.
