+++ helpers/jwtAuthCheck.js
