+++ routes/auth.js
@@ -4,16 +4,9 @@ const passport = require("passport");
 const mongoose = require('mongoose');
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcrypt');
-const nodemailer = require("nodemailer");
 const User = require('../models/user.js')
-
-const mailTransport = nodemailer.createTransport({
-	service: "gmail",
-	auth: {
-		user: process.env.NODE_MAILER_EMAIL,
-		pass: process.env.NODE_MAILER_PASSWORD
-	}
-});
+const sendVerificationEmail = require('../helpers/sendVerificationEmail');
+const sendForgetPasswordEmail = require('../helpers/sendForgetPasswordEmail');
 
 /* GET Google Authentication API. */
 router.get(
@@ -47,7 +40,10 @@ router.post("/login", (req, res) => {
 			if (bcrypt.compareSync(password, user.password)) {
 
 				const jwtToken = jwt.sign(
-					{ sub: user.id },
+					{
+						sub: user.id,
+						emailVerified: user.emailVerified
+					},
 					process.env.JWT_SECRET,
 					// { expiresIn: '30s' }
 				);
@@ -58,7 +54,8 @@ router.post("/login", (req, res) => {
 						user: {
 							email: user.email,
 							username: user.username,
-							organization: user.organization
+							organization: user.organization,
+							emailVerified: user.emailVerified
 						},
 					},
 					"token": jwtToken
@@ -100,7 +97,6 @@ router.post("/register", (req, res) => {
 	} else {
 
 		User.find({ $or: [{ 'email': email }, { 'username': username }] }).then(users => {
-			console.log(users);
 			if (users.length > 0) {
 
 				// Checking for duplication:
@@ -133,23 +129,35 @@ router.post("/register", (req, res) => {
 						newUser.save()
 							.then(user => {
 
-								const jwtToken = jwt.sign(
-									{ sub: user.id },
+								const accessToken = jwt.sign(
+									{
+										sub: user.id,
+										emailVerified: user.emailVerified
+									},
 									process.env.JWT_SECRET,
 									// { expiresIn: '30s' }
 								);
 
-								res.status(200).json({
-									"success": true,
-									"data": {
-										user: {
-											email: user.email,
-											username: user.username,
-											organization: user.organization
-										},
-									},
-									"token": jwtToken
-								})
+								sendVerificationEmail(user)
+									.then(response => {
+										console.log(`>>> Email Verification link sent to: ${user.email}`);
+										res.status(200).json({
+											"success": true,
+											"data": {
+												user: {
+													email: user.email,
+													username: user.username,
+													organization: user.organization,
+													emailVerified: user.emailVerified
+												},
+											},
+											"token": accessToken
+										})
+									})
+									.catch(err => {
+										console.error(err);
+									});
+
 							})
 							.catch(err => console.err(err))
 
@@ -165,77 +173,9 @@ router.post("/forgetPassword", (req, res) => {
 
 	User.findOne({ email: email }).then(user => {
 		if (user) {
-
-			// Generate token using userId:
-			const resetPasswordToken = jwt.sign(
-				{ sub: user.id },
-				process.env.JWT_SECRET,
-				{ expiresIn: '10m' } // Reset Link will be valid for only 5 mins.
-			);
-
-			// Send an email with generated Token:
-			const mailOptions = {
-				from: `"Candids" <${process.env.NODE_MAILER_EMAIL}>`,
-				to: email,
-				subject: "Candids - Reset Password",
-				text: "Candids - Reset Password",
-				// html: `Click here to reset your password ${resetPasswordToken}`
-				html: `
-				<div
-					class="email-background"
-					style="background: whitesmoke;font-family: sans-serif;text-align: center;font-size: 15px;padding: 14px 18px;"
-				>
-					<a
-						href="https://localhost:8080"
-						target="_blank"
-						rel="noopener noreferrer"
-						class="email-header"
-						style="text-decoration: none;color: #606060;"
-					>
-						<div class="header" style="font-weight: 800;font-size: 1.5em;">
-						Candids
-						</div>
-						<div class="subHeader" style="font-size: 1.1em;">
-						Sharing fun Moments
-						</div>
-					</a>
-
-					<div
-						class="email-container"
-						style="max-width: 650px;margin: 16px auto;background: white;padding: 32px;text-align: center;border-radius: 5px;"
-					>
-						<h1 style="font-size: 1.7em;margin-bottom: 1em;color: #323232;">
-						Reset Password
-						</h1>
-						<p style="line-height: 1.2em;color: #505050;">
-						Hey ${user.username}! We got a request to reset your Candids password.
-						</p>
-
-						<a
-						class="cta"
-						href="https://localhost:8080/resetPassword/?token=${resetPasswordToken}"
-						style="color: white;font-weight: 500;transition: all 0.2s ease-in-out;text-decoration: none;background: #0fc96c;display: inline-block;font-size: 1.2em;padding: 0.7em 1em;border-radius: 10em;margin: 1em 0;box-shadow: 0px 3px 6px rgba(0, 0, 0, 0.6);line-height: 1em;"
-						>Choose a new password</a
-						>
-						<p style="line-height: 1.2em;color: #505050;">
-						This link is only valid for 10 minutes.
-						</p>
-
-						<hr style="border: 0;border-top: 2px solid whitesmoke;margin: 30px;" />
-						<p
-						class="note"
-						style="line-height: 1.2em;color: #505050;margin-top: 3em;font-size: 0.9em;"
-						>
-						NOTE: If this wasn't you, you can simply ignore this email.
-						</p>
-					</div>
-				</div>`
-			};
-
-			mailTransport
-				.sendMail(mailOptions)
+			sendForgetPasswordEmail(user)
 				.then(response => {
-					console.log(`Password Reset link sent to: ${email}`);
+					console.log(`>>> Password Reset link sent to: ${user.email}`);
 					res.status(200).json({
 						"success": true,
 					});
@@ -246,7 +186,6 @@ router.post("/forgetPassword", (req, res) => {
 						"success": false,
 					});
 				});
-
 		} else {
 			res.status(400).send("Cannot find user")
 		}
@@ -263,37 +202,98 @@ router.post("/resetPassword", (req, res) => {
 				"success": false
 			});
 		else {
+			if (payload.type === 'reset') {
 
-			bcrypt.hash(password, +process.env.BCRYPT_SALT_ROUNDS)
-				.then(hashedPassword => {
+				bcrypt.hash(password, +process.env.BCRYPT_SALT_ROUNDS)
+					.then(hashedPassword => {
 
-					User.findByIdAndUpdate(payload.sub, {
-						password: hashedPassword
-					}).then(user => {
-						const jwtToken = jwt.sign(
-							{ sub: user.id },
-							process.env.JWT_SECRET,
-							// { expiresIn: '30s' }
-						);
-
-						res.status(200).json({
-							"success": true,
-							"data": {
-								user: {
-									email: user.email,
-									username: user.username,
-									organization: user.organization
+						User.findByIdAndUpdate(payload.sub, {
+							password: hashedPassword
+						}).then(user => {
+							const accessToken = jwt.sign(
+								{
+									sub: user.id,
+									emailVerified: user.emailVerified
 								},
-							},
-							"token": jwtToken
-						})
-					}).catch(err => console.err(err))
+								process.env.JWT_SECRET,
+								// { expiresIn: '30s' }
+							);
 
-				}); // Bcrypt
+							res.status(200).json({
+								"success": true,
+								"data": {
+									user: {
+										email: user.email,
+										username: user.username,
+										organization: user.organization,
+										emailVerified: user.emailVerified
+									},
+								},
+								"token": accessToken
+							})
+						}).catch(err => console.err(err))
 
-		} // Else
+					});
+
+			} else {
+				res.status(403).json({
+					"success": false
+				});
+			}
+		}
 	})
 })
 
+router.post("/verifyEmail", (req, res) => {
+	const { token } = req.body;
+
+	// Verify Token:
+	jwt.verify(token, process.env.JWT_SECRET, (err, payload) => {
+		if (err)
+			res.status(403).json({
+				"success": false
+			});
+		else {
+			if (payload.type === 'verification') {
+
+				bcrypt.hash(password, +process.env.BCRYPT_SALT_ROUNDS)
+					.then(hashedPassword => {
+
+						User.findByIdAndUpdate(payload.sub, {
+							password: hashedPassword
+						}).then(user => {
+							const accessToken = jwt.sign(
+								{
+									sub: user.id,
+									emailVerified: user.emailVerified
+								},
+								process.env.JWT_SECRET,
+								// { expiresIn: '30s' }
+							);
+
+							res.status(200).json({
+								"success": true,
+								"data": {
+									user: {
+										email: user.email,
+										username: user.username,
+										organization: user.organization,
+										emailVerified: user.emailVerified
+									},
+								},
+								"token": accessToken
+							})
+						}).catch(err => console.err(err))
+
+					});
+
+			} else {
+				res.status(403).json({
+					"success": false
+				});
+			}
+		}
+	})
+})
 
 module.exports = router;
